refactor: make a class for handling gameplay related to levels
